requirements.txt
django-crispy-forms
crispy-bootstrap5
django
psycopg2-binary
python-dotenv
whitenoise
paho-mqtt
xhtml2pdf
djangorestframework